Fixed rendering of pixels grid in PDF
### fixes
- Fixed rendering of pixels grid
- Added Roboto font
- Improved info page

+++ src/utils/file_storage.helper.ts
@@ -22,7 +22,8 @@ export class FileStorageHelper {
             break;
         }
 
-        await unlink(path);
+        if (file) await unlink(path);
+        
         return file;
     }
